+++ packages/polaris-middlewares/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.22](https://github.com/Enigmatis/polaris-middlewares/compare/@enigmatis/polaris-middlewares@1.4.21...@enigmatis/polaris-middlewares@1.4.22) (2020-09-10)
+
+
+### Bug Fixes
+
+* fixed permission directive definition ([#13](https://github.com/Enigmatis/polaris-middlewares/issues/13)) ([001f1bb](https://github.com/Enigmatis/polaris-middlewares/commit/001f1bb45c8c88fe4fef5c9308e9f35375d59245))
+
+
+
+
+
+## [1.4.21](https://github.com/Enigmatis/polaris-middlewares/compare/@enigmatis/polaris-middlewares@1.4.20...@enigmatis/polaris-middlewares@1.4.21) (2020-09-09)
+
+**Note:** Version bump only for package @enigmatis/polaris-middlewares
+
+
+
+
+
 ## [1.4.20](https://github.com/Enigmatis/polaris-middlewares/compare/@enigmatis/polaris-middlewares@1.4.19...@enigmatis/polaris-middlewares@1.4.20) (2020-09-07)